# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MarketShield AI - Financial Market Surveillance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trades with optional filters
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: riskLevel
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: suspicious
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeList"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Record a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get trade details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
...
components:
  schemas:
...
    Trade:
      type: object
      required: [id, symbol, traderName, tradeType, quantity, price, totalValue, riskScore, isSuspicious, timestamp, companyName, riskLevel]
      properties:
        id:
          type: integer
        symbol:
          type: string
        traderName:
          type: string
        tradeType:
          type: string
          enum: [buy, sell, short]
        quantity:
          type: integer
        price:
          type: number
        totalValue:
          type: number
        riskScore:
          type: number
        isSuspicious:
          type: boolean
        timestamp:
          type: string
        companyName:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        anomalyFlags:
          type: ["array", "null"]
          items:
            type: string
    TradeInput:
      type: object
      required: [symbol, traderName, tradeType, quantity, price]
      properties:
        symbol:
          type: string
        traderName:
          type: string
        tradeType:
          type: string
          enum: [buy, sell, short]
        quantity:
          type: integer
        price:
          type: number
    TradeList:
      type: object
      required: [trades, total]
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
...